:construction: Desenvolvendo as orders da parte do produtor

+++ admin-app/src/app/pages/home/home.component.ts
@@ -1,4 +1,4 @@
-import { Component, HostListener, OnInit } from '@angular/core';
+import { Component, ElementRef, HostListener, OnInit, ViewChild } from '@angular/core';
 import { ProdutoService } from '../../autentication/service/produto/produto.service';
 import { PaginationComponent } from "../../shared/_component/pagination/pagination.component";
 import { CurrencyPipe } from '@angular/common';
@@ -7,19 +7,18 @@ import { FormsModule } from '@angular/forms';
 import { debounceTime, distinctUntilChanged, Subject } from 'rxjs';
 import { AlertComponent } from '../../shared/models/alert/alert.component';
 
-@Component({ 
+@Component({
   selector: 'app-home',
   standalone: true,
-  imports: [PaginationComponent, AlertComponent, CurrencyPipe, RouterLink, RouterLinkActive, FormsModule],
+  imports: [PaginationComponent, AlertComponent, CurrencyPipe, RouterLink, FormsModule],
   templateUrl: './home.component.html',
   styleUrl: './home.component.css'
 })
 export class HomeComponent implements OnInit {
-  isDropdownOpen = false;
   produtos: { id: number; productName: string; price: number; quantity: number; description: string }[] = [];
 
-  itemsPerPage = 20; // Quantos produtos por página
-  currentPage = 1; // Página inicial
+  itemsPerPage = 20;
+  currentPage = 1; 
   isMenuOpen = false;
 
   showAlert: boolean = false;
@@ -30,68 +29,58 @@ export class HomeComponent implements OnInit {
   isSearching: boolean = false;
   private searchSubject = new Subject<string>();
 
+
+  isDeleteModalOpen: boolean = false;
+  isInfoModalOpen: boolean = false;
+  selectedProductId: number | null = null;
+  selectedProductsIds: number[] = [];
+
   constructor(
     private produtosService: ProdutoService,
     private router: Router
-  ) {}
+  ) { }
 
   ngOnInit(): void {
     this.listProducts();
 
-    // Configura busca com debounce
     this.searchSubject.pipe(
-      debounceTime(300), // Aguarda 300ms antes de buscar
-      distinctUntilChanged() // Evita buscas duplicadas
+      debounceTime(300), 
+      distinctUntilChanged() 
     ).subscribe(term => {
       if (term.trim() === '') {
-        this.listProducts(); // Se estiver vazio, carrega tudo novamente
+        this.listProducts(); 
       } else {
         this.getByNameProduct(term);
       }
     });
   }
 
-  // Total de páginas calculado com base no número de produtos
   get totalPages() {
     return Math.ceil(this.produtos.length / this.itemsPerPage);
   }
 
-  // Filtra os produtos de acordo com a página atual
   get paginatedProdutos() {
     const startIndex = (this.currentPage - 1) * this.itemsPerPage;
     return this.produtos.slice(startIndex, startIndex + this.itemsPerPage);
   }
 
-  // Atualiza a página atual (chamado pelo componente de paginação)
   onPageChange(page: number): void {
-    this.currentPage = page; // Atualiza a página atual
-  }
-
-  toggleDropdown3() {
-    this.isDropdownOpen = !this.isDropdownOpen;
-  }
-
-  @HostListener('document:click', ['$event'])
-  closeDropdown(event: Event) {
-    const target = event.target as HTMLElement;
-    if (!target.closest('.dropdown-container')) {
-      this.isDropdownOpen = false;
-    }
+    this.currentPage = page; 
   }
 
   listProducts(): void {
-      this.produtosService.getProdutos().subscribe({
-        next: (dados) => {
-          this.produtos = dados;
-        },
-        error: (err) => {
-          console.error('Erro ao buscar produtos:', err);
-        },
-      });
+    this.produtosService.getProdutos().subscribe({
+      next: (dados) => {
+        this.produtos = dados;
+      },
+      error: (err) => {
+        console.error('Erro ao buscar produtos:', err);
+      },
+    });
   }
 
   buscarProdutos(): void {
-    this.searchSubject.next(this.searchTerm); // Envia o termo para ser processado
+    this.searchSubject.next(this.searchTerm); 
   }
 
   getByNameProduct(nome: string): void {
@@ -115,21 +104,58 @@ export class HomeComponent implements OnInit {
     );
   }
 
-  deleteProduct(id: number): void {
-    this.produtosService.deleteProduct(id).subscribe(
-      () => {
-        this.listProducts();
-        this.showAlert = true;
-        this.message = "Produto deletado com sucesso.";
-        this.categAlert = 3;
+  selectProducts(id: number): void {
+    this.selectedProductId = id;
+    setTimeout(() => {
+      this.selectedProductId = null;
+    }, 5000);
+    const index = this.selectedProductsIds.indexOf(id);
+    if (index === -1) {
+      this.selectedProductsIds.push(id); 
+    } else {
+      this.selectedProductsIds.splice(index, 1); 
+    }
+  }
 
-      },
-      error => {
-        this.showAlert = true;
-        this.message = "Erro ao deletar produto.";
-        this.categAlert = 2;
-      }
-    );
+  deleteProduct(): void {
+    if (this.selectedProductsIds.length > 0) {
+      this.selectedProductsIds.forEach(id => {
+        this.produtosService.deleteProduct(id).subscribe(
+          () => {
+            this.listProducts(); 
+            this.showAlert = true;
+            this.message = `Produto com ID ${id} deletado com sucesso.`;
+            this.categAlert = 3;
+            console.log('Deletando o produto com id:', id);
+          },
+          error => {
+            this.showAlert = true;
+            this.message = "Erro ao deletar produto.";
+            this.categAlert = 2;
+          }
+        );
+      });
+      this.selectedProductsIds = []; // Limpa a seleção após a exclusão
+    } else {
+      this.showAlert = true;
+      this.message = "Nenhum produto selecionado.";
+      this.categAlert = 2;
+    }
+  }
+
+  showDeleteModal() {
+    this.isDeleteModalOpen = true;
+    this.isInfoModalOpen = false; 
+  }
+
+  showInfoModal() {
+    this.isInfoModalOpen = true;
+    this.isDeleteModalOpen = false; 
+  }
+
+  closeModal() {
+    this.isDeleteModalOpen = false;
+    this.isInfoModalOpen = false;
   }
 
 }
